# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/trocas": {
            "post": {
                "description": "Registra uma proposta de troca com base nos dados enviados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trocas"
                ],
                "summary": "Cadastra nova proposta",
                "parameters": [
                    {
                        "description": "Dados da proposta",
                        "name": "proposta",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/propostarepo.PropostaWriteModel"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Proposta criada com sucesso e ID retornado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Body inválido",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro ao salvar proposta",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/trocas/{id}/historico": {
            "get": {
                "description": "Retorna propostas enviadas ou recebidas por um usuário com base no tipo e status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trocas"
                ],
                "summary": "Lista propostas do usuário",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do usuário",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tipo de proposta (enviadas ou recebidas)",
                        "name": "tipo",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Status da proposta (pendente, aceita, recusada)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/propostarepo.PropostaFormatada"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/trocas/{id}/status": {
            "put": {
                "description": "Altera o status da proposta (aceita, recusada, pendente)",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "trocas"
                ],
                "summary": "Atualiza o status da proposta",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da proposta",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Novo status",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/propostahandler.StatusUpdatePayload"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Atualizado com sucesso",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "propostahandler.StatusUpdatePayload": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "propostarepo.Produto": {
            "type": "object",
            "properties": {
                "categoria": {
                    "type": "string"
                },
                "descricao": {
                    "type": "string"
                },
                "imagens": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "nome": {
                    "type": "string"
                },
                "usuario": {
                    "type": "string"
                }
            }
        },
        "propostarepo.PropostaFormatada": {
            "type": "object",
            "properties": {
                "produto_postagem": {
                    "$ref": "#/definitions/propostarepo.Produto"
                },
                "produto_proposta_troca": {
                    "$ref": "#/definitions/propostarepo.Produto"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "propostarepo.PropostaWriteModel": {
            "type": "object",
            "properties": {
                "categoria": {
                    "type": "string"
                },
                "descricao": {
                    "type": "string"
                },
                "dono_postagem_id": {
                    "type": "string"
                },
                "interessado_id": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "postagem_id": {
                    "type": "string"
                }
            }
        },
...
    }
}